+++ src/main/java/com/braintreegateway/WebhookTestingGateway.java
@@ -582,8 +582,11 @@ public class WebhookTestingGateway {
 
     private String localPaymentCompletedXml() {
         return node("local-payment",
-                node("payment-id", "a-payment-id"),
+                node("bic", "a-bic"),
+                node("iban-last-chars", "1234"),
                 node("payer-id", "a-payer-id"),
+                node("payer-name", "a-payer-name"),
+                node("payment-id", "a-payment-id"),
                 node("payment-method-nonce", "ee257d98-de40-47e8-96b3-a6954ea7a9a4"),
                 node("transaction",
                      node("id", "1"),