+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Unreleased
+## 3.36.0
+* Add `payerName`, `bic` and `iban_last_chars` for LocalPaymentCompleted webhook
+* Add `editPaypalVaultId` to `PayPalAccount`
 * Add `aniFirstNameResponseCode` and `aniLastNameResponseCode` to `CreditCardVerification`
 * Add `shippingTaxAmount` to `Transaction` and `TransactionRequest`